# openapi/ewoh.yaml
openapi: 3.0.3
info:
  title: EWOH Embodied Factory OS API
  version: 1.0.0
  description: >-
    OpenAPI contract for EWOH NestJS controllers. Auth, dashboard, world, task, resource, alert, approval, control, AI,
    scheduler, gamification, ingest, files, simulator, spatial, system, audit, and health surfaces.
servers:
  - url: https://api.ewoh.local
    description: Production/edge gateway
  - url: http://127.0.0.1:3000
    description: Local NestJS dev server
security:
  - BearerAuth: []
...
paths:
...
  /api/devices:
    get:
      tags:
        - devices
      summary: List devices
      parameters: &ref_0
        - &ref_1
          name: keyword
          in: query
          required: false
          description: Search by device id, worker name, or model
          schema:
            type: string
        - &ref_2
          name: online
          in: query
          required: false
          description: Filter by online state
          schema:
            type: boolean
        - &ref_3
          name: batteryMin
          in: query
          required: false
          description: Minimum battery percentage
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - &ref_4
          name: batteryMax
          in: query
          required: false
          description: Maximum battery percentage
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - &ref_5
          name: sourceType
          in: query
          required: false
          description: real, controlled_test, simulated, replayed, stale, or offline
          schema:
            type: string
        - &ref_6
          name: model
          in: query
          required: false
          description: Device model
          schema:
            type: string
        - &ref_7
          name: orderby
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum:
              - battery
              - batteryDesc
              - lastTelemetryAt
              - lastTelemetryAtDesc
              - deviceId
              - deviceIdDesc
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceList'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/devices/{id}:
    get:
      tags:
        - devices
      summary: Device detail
      parameters:
        - name: id
          in: path
          required: true
          description: Device id
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceInfo'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/devices/{id}/bindings:
    post:
      tags:
        - devices
      summary: Bind device to spatial entity or person
      parameters:
        - name: id
          in: path
          required: true
          description: Device id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindDeviceRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceBinding'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        InternalError:
          $ref: '#/components/responses/InternalError'
...
  /api/dashboard/devices:
    get:
      tags:
        - devices
      summary: List dashboard devices
      parameters: *ref_0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceList'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - devices
      summary: Create device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeviceDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceInfo'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
...
  /api/dashboard/devices/{deviceId}:
    patch:
      tags:
        - devices
      summary: Update device
      parameters:
        - name: deviceId
          in: path
          required: true
          description: Device id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeviceDto'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceInfo'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/dashboard/devices/{deviceId}/bindings:
    get:
      tags:
        - devices
      summary: Get device bindings
      parameters:
        - name: deviceId
          in: path
          required: true
          description: Device id
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceBinding'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - devices
      summary: Bind dashboard device
      parameters:
        - name: deviceId
          in: path
          required: true
          description: Device id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindDeviceRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceBinding'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
        - devices
      summary: Unbind dashboard device
      parameters:
        - name: deviceId
          in: path
          required: true
          description: Device id
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
...
  /api/mes/work-orders:
    get:
      tags:
        - mes
      summary: List MES work orders
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MesWorkOrderList'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - mes
      summary: Create MES work order with production steps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMesWorkOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MesWorkOrderDetail'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/mes/work-orders/{id}:
    get:
      tags:
        - mes
      summary: MES work order detail with steps and materials
      parameters:
        - name: id
          in: path
          required: true
          description: Work order id
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MesWorkOrderDetail'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/mes/work-orders/{id}/trace:
    get:
      tags:
        - mes
      summary: Quality trace graph for a work order
      parameters:
        - name: id
          in: path
          required: true
          description: Work order id
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MesTraceGraph'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/mes/work-orders/{id}/state:
    post:
      tags:
        - mes
      summary: MES work order state transition
      parameters:
        - name: id
          in: path
          required: true
          description: Work order id
          schema:
            type: string
        - name: action
          in: query
          required: true
          description: release, start, complete, or cancel
          schema:
            type: string
            enum:
              - release
              - start
              - complete
              - cancel
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MesWorkOrder'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        Conflict:
          $ref: '#/components/responses/Conflict'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/mes/work-orders/{id}/steps/{stepId}/state:
    post:
      tags:
        - mes
      summary: MES step state transition
      parameters:
        - name: id
          in: path
          required: true
          description: Work order id
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          description: Step id
          schema:
            type: string
        - name: action
          in: query
          required: true
          description: start, report, review, handover, pause, resume, or cancel
          schema:
            type: string
            enum:
              - start
              - report
              - review
              - handover
              - pause
              - resume
              - cancel
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MesStepTransitionRequest'
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MesStep'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        Conflict:
          $ref: '#/components/responses/Conflict'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/mes/work-orders/{id}/materials:
    get:
      tags:
        - mes
      summary: List material consumption for a work order
      parameters:
        - name: id
          in: path
          required: true
          description: Work order id
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialConsumptionList'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - mes
      summary: Record material consumption for a work order
      parameters:
        - name: id
          in: path
          required: true
          description: Work order id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialConsumptionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialConsumption'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/mes/work-orders/{id}/inspections:
    post:
      tags:
        - mes
      summary: Record a quality inspection for a work order step
      parameters:
        - name: id
          in: path
          required: true
          description: Work order id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QualityInspectionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityInspectionResult'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    IngestKey:
      type: apiKey
      in: header
      name: X-Ingest-Key
...
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: State or conflict error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Legacy business or validation error code
            errorCode:
              type: string
              description: Unified business or validation error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: string
              description: Optional structured details
            fieldErrors:
              type: object
              description: Field-level validation errors
              additionalProperties:
                type: array
                items:
                  type: string
            requestId:
              type: string
              description: Request ID used to correlate logs, audit, and support bundles
            retryable:
              type: boolean
              description: Whether the operation can be safely retried
            recommendedAction:
              type: string
              description: Operator-facing recommended next action
            timestamp:
              type: number
              description: Error timestamp in milliseconds
          required:
            - code
            - errorCode
            - message
            - requestId
            - retryable
            - recommendedAction
      required:
        - error
      additionalProperties: false
...
    DeviceInfo:
      type: object
      properties:
        id:
          type: string
          description: Database row id
        deviceId:
          type: string
          description: External device id
        workerName:
          type: string
          description: Bound worker display name
        deviceModel:
          type: string
          description: Device model
        batteryPct:
          type: number
          description: Battery percentage 0-100
        online:
          type: boolean
          description: Current online state
        lastTelemetryAt:
          type: string
          nullable: true
        entityId:
          type: string
          description: Linked spatial entity id
        parentId:
          type: string
          nullable: true
        x:
          type: number
          description: Spatial x coordinate in mm
        y:
          type: number
          description: Spatial y coordinate in mm
        sourceType:
          type: string
          description: real, controlled_test, simulated, replayed, stale, or offline
        firmwareVersion:
          type: string
          nullable: true
        hardwareVersion:
          type: string
          nullable: true
        protocolVersion:
          type: string
          nullable: true
        temperatureC:
          type: number
          nullable: true
        faultCode:
          type: string
          nullable: true
        lastRawRef:
          type: string
          nullable: true
        boundPersonId:
          type: string
          nullable: true
        boundPersonName:
          type: string
          nullable: true
      required: []
      additionalProperties: false
    DeviceList:
      type: array
      items:
        $ref: '#/components/schemas/DeviceInfo'
...
    CreateDeviceDto:
      type: object
      properties:
        deviceId:
          type: string
          description: Unique external device id
        workerName:
          type: string
          description: Worker display name
        deviceModel:
          type: string
          description: Device model
        batteryPct:
          type: number
          description: Initial battery percentage 0-100
        online:
          type: boolean
          description: Initial online state
        sourceType:
          type: string
          description: real, controlled_test, simulated, replayed, stale, or offline
        firmwareVersion:
          type: string
          description: Firmware version
        hardwareVersion:
          type: string
          description: Hardware version
        protocolVersion:
          type: string
          description: Protocol version
      required:
        - deviceId
      additionalProperties: false
    UpdateDeviceDto:
      type: object
      properties:
        workerName:
          type: string
          description: Worker display name
        deviceModel:
          type: string
          description: Device model
        batteryPct:
          type: number
          description: Battery percentage 0-100
        online:
          type: boolean
          description: Online state
        faultCode:
          type: string
          description: Fault code
        firmwareVersion:
          type: string
          description: Firmware version
        hardwareVersion:
          type: string
          description: Hardware version
        protocolVersion:
          type: string
          description: Protocol version
        temperatureC:
          type: number
          description: Device temperature in Celsius
      required: []
      additionalProperties: false
    BindDeviceRequest:
      type: object
      properties:
        spatialEntityId:
          type: string
          nullable: true
          description: Target spatial entity id
        personEntityId:
          type: string
          nullable: true
          description: Person entity id
      required: []
      additionalProperties: false
    DeviceBinding:
      type: object
      properties:
        deviceId:
          type: string
          description: Bound device id
        spatialEntityId:
          type: string
          nullable: true
        hierarchyPath:
          type: array
          items:
            type: object
            properties:
              entityId:
                type: string
                description: Spatial entity id
              name:
                type: string
                description: Entity name
              entityType:
                type: string
                description: Entity type
            required: []
            additionalProperties: false
        boundPersonId:
          type: string
          nullable: true
        boundPersonName:
          type: string
          nullable: true
      required: []
      additionalProperties: false
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
      additionalProperties: false
...
    MesTraceGraph:
      type: object
      properties:
        workOrder:
          $ref: '#/components/schemas/MesWorkOrder'
        steps:
          $ref: '#/components/schemas/MesStepList'
        materials:
          $ref: '#/components/schemas/MaterialConsumptionList'
        inspections:
          type: array
          items:
            type: object
        nodes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
              label:
                type: string
        links:
          type: array
          items:
            type: object
            properties:
              from:
                type: string
              to:
                type: string
              type:
                type: string
...
    MesWorkOrder:
      type: object
      properties:
        scheduleTaskId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - draft
            - released
            - in_progress
            - completed
            - cancelled
        priority:
          type: string
        source:
          type: string
        planStart:
          type: string
          nullable: true
        planEnd:
          type: string
          nullable: true
        actualStart:
          type: string
          nullable: true
        actualEnd:
          type: string
          nullable: true
        progress:
          type: integer
        isSimulation:
          type: boolean
    MesWorkOrderList:
      type: array
      items:
        $ref: '#/components/schemas/MesWorkOrder'
    MesStep:
      type: object
      properties:
        stepId:
          type: string
        scheduleTaskId:
          type: string
        stepNo:
          type: integer
        name:
          type: string
        instruction:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - pending
            - in_progress
            - paused
            - reported
            - reviewed
            - handed_over
            - cancelled
        actualStart:
          type: string
          nullable: true
        actualEnd:
          type: string
          nullable: true
        assignedPersonId:
          type: string
          nullable: true
        assignedDeviceId:
          type: string
          nullable: true
        spatialEntityId:
          type: string
          nullable: true
        progress:
          type: integer
        resultJson:
          type: object
          nullable: true
    MesStepList:
      type: array
      items:
        $ref: '#/components/schemas/MesStep'
    MesWorkOrderDetail:
      type: object
      properties:
        workOrder:
          $ref: '#/components/schemas/MesWorkOrder'
        steps:
          $ref: '#/components/schemas/MesStepList'
        materials:
          $ref: '#/components/schemas/MaterialConsumptionList'
    MesStepInput:
      type: object
      properties:
        name:
          type: string
        instruction:
          type: string
        assignedPersonId:
          type: string
        assignedDeviceId:
          type: string
        spatialEntityId:
          type: string
        plannedStart:
          type: string
        plannedEnd:
          type: string
      required:
        - name
    CreateMesWorkOrderRequest:
      type: object
      properties:
        orderId:
          type: string
        title:
          type: string
        productCode:
          type: string
        orderQty:
          type: number
        batchNo:
          type: string
        priority:
          type: string
        planStart:
          type: string
        planEnd:
          type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/MesStepInput'
      required:
        - title
        - steps
    MesStepTransitionRequest:
      type: object
      properties:
        quantity:
          type: number
        note:
          type: string
        decision:
          type: string
        receiver:
          type: string
        operatorId:
          type: string
        reviewer:
          type: string
    MaterialConsumption:
      type: object
      properties:
        bindingId:
          type: string
        bindingType:
          type: string
        resourceType:
          type: string
        resourceId:
          type: string
        targetType:
          type: string
        targetId:
          type: string
        quantity:
          type: number
        reason:
          type: string
          nullable: true
        status:
          type: string
        operatorId:
          type: string
          nullable: true
        startTime:
          type: string
    MaterialConsumptionList:
      type: array
      items:
        $ref: '#/components/schemas/MaterialConsumption'
    MaterialConsumptionRequest:
      type: object
      properties:
        materialId:
          type: string
        quantity:
          type: number
        reason:
          type: string
        operatorId:
          type: string
      required:
        - materialId
        - quantity
    QualityInspectionRequest:
      type: object
      properties:
        stepId:
          type: string
        inspectorId:
          type: string
        result:
          type: string
          enum:
            - pass
            - fail
            - rework
        defectCode:
          type: string
        quantity:
          type: number
        note:
          type: string
        schemeId:
          type: string
        stage:
          type: string
          enum: [first, in_process, final]
        checkResults:
          type: array
          items:
            type: object
            required:
              - itemId
              - result
            properties:
              itemId:
                type: string
              result:
                type: string
                enum: [pass, fail]
              note:
                type: string
      required:
        - stepId
        - result
    QualityInspectionResult:
      type: object
      properties:
        stepId:
          type: string
        eventId:
          type: string
        result:
          type: string
...